feat: subcommands handling
