+++ src/commands/tools/custombot/index.js
@@ -1,0 +1,4 @@
+import { SlashCommandBuilder } from 'discord.js';
+export const data = new SlashCommandBuilder()
+  .setName('custombot')
+  .setDescription('custombot')
